fix(input): invalid font declaration (#5154)
Fixes the `font` declaration on the `md-input-container` being ignored, because it was invalid.

Note: I updated the font size to the correct value from [the spec](https://material.io/guidelines/components/text-fields.html#text-fields-states), however I'm not using it, in order to avoid any breaking changes. Also note that since the line height wasn't being used before, the baseline is now slightly off when next to a select.

Fixes #5262.