chore: fix warning about unconfigurable property (#9599)
Fixes that some browsers in Browserstack or Saucelabs warn about the `defineProperty` call in the interactivity checker tests.

`LOG: 'Attempting to configure 'frameElement' with descriptor '{}' on object '[object Window]' and got error, giving up: TypeError: Attempting to change the getter of an unconfigurable property.'`.

This happens because some browsers explicitly set all properties to `configurable: false` on `window` objects. Since this is only happening inside of Safari, we don't overwrite the property.