fix(overlay): explicitly implement OverlayReference. (#11824)
`OverlayRef` structurally matches `OverlayReference`, but does not
explicitly implement it. That works in TypeScript, and should work in
Closure Compiler through the use of `@record`, but for unclear reasons
doesn't.

Explicitly implementing the interface makes this code clearer, and works
around the problem in Closure Compiler while that is being fixed.