fix(menu): scroll position jumping to top after animation is done on scrollable menu (#14190)
Originally in #11859 we introduced some logic that makes sure that the menu panel is always scrolled to the top when it's open. This works, but is janky, because the user can see the menu being scrolled down while it's animating. These changes move the logic, that reset the scroll position, from the `done` callback to `start`.

Fixes #11790.