fix(layout): breakpoint observer completing unrelated subscribers when preceding subscriber unsubscribes (#14988)
Fixes the `BreakpointObserver` completing subscribers that come after the current one, if one of the preceding subscribers unsubscribes. The issue seems to come from the way we store the listener in order to remove it later on. These changes switch to creating the `Observable` ourselves, rather than going through `fromEventPattern` so that we have more control over the event handler.

Fixes #14983.