fix(slide-toggle): thumb animation not working on mobile (#15236)
With https://github.com/angular/material2/commit/cc8f871e8d3978cca95bffb3c0ab0d79961a8ddd, we enabled the `grabbing` cursor if the thumb container is pressed. This accidentally broke animation for the thumb sometimes because the `:active`
pseudo class is added on tap/touch and now accidentally sets the thumb transition duration to `0ms`. Rather we should only set the thumb transition duration to `0ms` if the user started dragging the thumb (with a small threshold)

Fixes #15232