build: fix examples package generating invalid metadata (#16073)
When building the `@angular/material-examples` package with
Bazel, the generated flat module metadata file does not include
metadata for the `ExampleModule`. This is because there is a
genrule which declares an output with the same name of the
example module source file. This means that the `ng_module`
rule somehow accidentally picks up the genrule output and the
generated flat module metadata file does not include the example
module due https://github.com/angular/angular/issues/30259