build: replace setup angular snapshots script with python script
In order to be able to take advantage of Google's official
bazel docker images, we no longer can use a NodeJS script
for setting up the snapshot builds because NodeJS is not
installed in the image. Rather we need to convert the script
to a python script in order to be able to perform the package.json
modification.

Since this script is not intended to be executed locally, we
are free to use Python for this.