Rework ng-update to no longer depend on tslint (#16433)
* fix(ng-update): no longer depend on tslint for migration

Reworks `ng update` to no longer depend on tslint for the
migration. TSlint caused various problems:

* No support for HTML templates and stylesheets (workaround was needed)
* Reruns all upgrade lint rules after file has been updated (significant performance issue for projects with a lot of files)
* Recreates the TypeScript program each time a source file has been updated (significant memory pressure for big TypeScript projects, causing OOM exceptions)
* TSLint recursively visits the nodes of all source files for each upgrade lint rule (performance issue)
* TSLint is not guaranteed to be installed in CLI projects. See: https://github.com/angular/angular-cli/issues/14555
* TSLint replacements lead to memory leaks due to the retained TypeScript nodes
* No way to have a *global analysis* phase since lint rules are only able to visit source files.
* No flexibility. i.e.
  * No way to ensure source files are only analyzed a single time
  * No way to implement a progress bar
  * No easy way to add support for HTML templates or stylesheets

* chore: reformat update schematics code with clang-format

Reformats the `ng update` code within the CDK and Material package
with clang-format. This ensures that the update code is consistent.

* docs: update schematic architecture document

Updates the schematic architecture document to reflect the
new state of the `ng update` implementation which no longer
uses `tslint` under the hood.

* perf(ng-update): do not run migration rules if there is no upgrade data

* No longer runs migration rules if there is no upgrade data. Running rules
conditionally is now easy to implement with the new `update-tool`.
* Adds comments for misc rule migrations to show which PR the given misc
migration covers.

* perf(ng-update): only migrate projects referenced in CLI workspace config

Currently `ng update` tries to resolve projects by common TypeScript tsconfig
locations besides looking in the CLI workspace config. This means that projects
which are technically not part of a workspace project are still migrated even
though they don't need to be. Similarly to `angular/angular`, we only migrate
projects explicitly referenced in the `angular.json` workspace configuration.

* refactor(ng-update): better messaging for migration failures

Previously with tslint, we had no way to determine whether there were
any upgrade rules which added rule failures without an automatic migration.

With the `update-tool` we can now conditionally print a better a message
if there were rule failures which need manual developer attention. i.e. we
only print the message that asks the developer to check for any failures in
the terminal output if there were actual rule failures.