feat(material-experimental): add test harness for radio-group
Follow-up for 583af19e9cc6fb37e1e4d32fe2abcf9700d12fff. This commit
introduces a new test harness for the `mat-radio-group` implementation.

Note that we can't provide harness methods for getting the selected
value because the selected value or the value of a radio-button are
part of the component-internal state (no way to determine from DOM)