fix(schematics): module imports not generated for schematics (#16787)
Currently in some scenarios the CLI project has a different TypeScript version
than the `typescript` version used in the `@schematics/angular` package.

This causes AST parsing & transformations to be broken due to the shifted
values (e.g. in the `SyntaxKind` enum). Previously we solved this by always
using the `typescript` version brought in by the `@schematics/angular` package,
but this no longer works as of v8.0.0 CLI because the `@schematics/angular`
package now vendors `typescript` instead of having an explicit version as
dependency.

https://github.com/angular/angular-cli/commit/bf1c069f73c8e3d4f0e8d584cbfb47c408c1730b