build: workaround for dev-app livereloading with ibazel (#16882)
Adds a workaround that makes livereloading work most of the time. It won't
work if the default port for the ibazel livereloading script has changed. This
can happen if the port is already used/reserved or if the default port changes
in the `livereload-server` implementation.

Ideally the livereloading script would be served from a canonical location
that always works (regardless of the port). A proposal for this has been
submitted within: https://github.com/bazelbuild/rules_typescript/pull/468