build: create custom bazel dev-server rule (#16937)
Implements a custom bazel dev-server rule that can be exposed eventually. The reason
we need a custom dev-server implementation is that the "ts_devserver" is not flexible
and needs to be synced into google3 (causing slow syncing; and hestitancy to adding
new features. always the question of scope).

We need our own implemenation because we want:

* Live-reloading to work (https://github.com/bazelbuild/rules_nodejs/issues/1036)
* HTML History API support (currently the ts_devserver always sends a 404 status code)
* Better host binding of the server (so that we can access the server on other devices)
* Flexibility & control (being able to do changes so that the dev-server fits our needs)