build: move edge browser test to saucelabs (#16985)
Browserstack Edge is currently very flaky and we don't seem
to be able to fix the flakiness on our end. Recording and debugging
live browserstack instances doesn't seem to help and the behavior
of the browser seems cached. e.g. commit that sets up the MDC
theme for the tests is not reflected in the BrowserStack browser video.

All the test failures are not reproducible locally in Edge, nor do they
reproduce if the tests are executed through BrowserStack from the
local machine (following the CI steps locally).

Since we don't want to spend too much time on it, and Edge runs
extremely slow and flaky on BrowserStack anyway.. we move it
over to Saucelabs.