fix(drag-drop): incorrectly constraining element position if size changes between drag sequences (#16876)
When a drag sequence is started we cache the dimensions of the element and the boundary so that we don't have to cause a reflow for each pixel of dragging, however with the current logic the cached dimensions are only cleared for the case where the item is part of a drop list. As a result, if a freely-dragged item with a boundary changes size between drag sequences its position won't be constrained correctly anymore.

These changes fix the issue by correctly clearing the cached dimensions in all cases.

Fixes #15749.