fix(drag-drop): unable to drop into connected sibling after scrolling into view via the parent (#17162)
In #16681 we added some logic that updates the cached dimensions if the document is scrolled while dragging, however it didn't account for any scrollable parent nodes that might've been scrolled. These changes switch the event listener to use capturing so that we can pick up all of the elements being scrolled.

Fixes #17144.