build: re-add theming scss bundle to material package (#17278)
Re-adds the `_theming.scss` bundle the to the Material
release package. We previously switched it to a simple
sass file that did not bundle all imported mixins.

We need to switch it back to a bundle since we observed
that importing individual files is slower than accessing
the bundled sass file.