build: ensure angular version placeholder is up-to-date when staging release (#17465)
Ensures that the Angular version placeholder is properly updated when
staging the release. We perform this as part of the staging script,
since changes between staging and publishing are *not* expected
and this is enforced by the publish script.

Additionally, the rules are very strict. If something ever changes
due to unknown reasons, the peer dependency checks needs to be
disabled manually / or updated.